# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface SocialLinks {
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  tiktok?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  linkedin?: string | null;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  socialLinks?: SocialLinks;
  /** @nullable */
  lastLogin?: string | null;
  /** @nullable */
  lastLogout?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  socialLinks?: SocialLinks;
}

export interface UpdateRoleBody {
  role: string;
}

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  createdBy: number;
  createdAt: string;
  updatedAt: string;
  lessonCount: number;
  enrollmentCount: number;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  content?: string | null;
  order: number;
  createdAt: string;
}

export interface CourseWithLessons {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  createdBy: number;
  createdAt: string;
  updatedAt: string;
  lessonCount: number;
  enrollmentCount: number;
  lessons: Lesson[];
  isEnrolled: boolean;
  /** @nullable */
  progress?: number | null;
}

export interface CreateCourseBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
}

export interface UpdateCourseBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
}

export interface CreateLessonBody {
  title: string;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  content?: string | null;
  order: number;
}

export interface UpdateLessonBody {
  title?: string;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  content?: string | null;
  order?: number;
}

export interface Enrollment {
  id: number;
  userId: number;
  courseId: number;
  progress: number;
  createdAt: string;
}

export interface EnrollmentWithCourse {
  id: number;
  userId: number;
  courseId: number;
  progress: number;
  createdAt: string;
  course: Course;
}

export interface EnrollBody {
  courseId: number;
}

export interface UpdateProgressBody {
  progress: number;
}

export interface PostWithAuthor {
  id: number;
  userId: number;
  content: string;
  likeCount: number;
  commentCount: number;
  isLiked: boolean;
  createdAt: string;
  author: User;
}

export interface CreatePostBody {
  content: string;
}

export interface LikeResponse {
  liked: boolean;
  likeCount: number;
}

export interface CommentWithAuthor {
  id: number;
  postId: number;
  userId: number;
  comment: string;
  createdAt: string;
  author: User;
}

export interface CreateCommentBody {
  comment: string;
}

export interface DashboardStats {
  totalUsers: number;
  totalCourses: number;
  totalEnrollments: number;
  totalPosts: number;
  myEnrollments: number;
  myProgress: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  message: string;
  createdAt: string;
  userId: number;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
}