# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseWithLessons"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Lessons
  /courses/{courseId}/lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List lessons for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create lesson
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Enrollments
  /enrollments:
    get:
      operationId: listMyEnrollments
      tags: [enrollments]
      summary: List current user enrollments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrollmentWithCourse"
    post:
      operationId: enrollInCourse
      tags: [enrollments]
      summary: Enroll in course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollBody"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{courseId}/progress:
    patch:
      operationId: updateProgress
      tags: [enrollments]
      summary: Update lesson progress
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgressBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"

  # Posts
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SignupBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
        lastLogin:
          type: ["string", "null"]
        lastLogout:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, name, role, createdAt]
    SocialLinks:
      type: object
      properties:
        facebook:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        tiktok:
          type: ["string", "null"]
        twitter:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        createdBy:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        lessonCount:
          type: integer
        enrollmentCount:
          type: integer
      required: [id, title, createdBy, createdAt, updatedAt, lessonCount, enrollmentCount]
    CourseWithLessons:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        createdBy:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        lessonCount:
          type: integer
        enrollmentCount:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
        isEnrolled:
          type: boolean
        progress:
          type: ["integer", "null"]
      required: [id, title, createdBy, createdAt, updatedAt, lessonCount, enrollmentCount, lessons, isEnrolled]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
      required: [title]
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
    Lesson:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        videoUrl:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required: [id, courseId, title, order, createdAt]
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        videoUrl:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        order:
          type: integer
      required: [title, order]
    UpdateLessonBody:
      type: object
      properties:
        title:
          type: string
        videoUrl:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        order:
          type: integer
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        progress:
          type: integer
        createdAt:
          type: string
      required: [id, userId, courseId, progress, createdAt]
    EnrollmentWithCourse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        progress:
          type: integer
        createdAt:
          type: string
        course:
          $ref: "#/components/schemas/Course"
      required: [id, userId, courseId, progress, createdAt, course]
    EnrollBody:
      type: object
      properties:
        courseId:
          type: integer
      required: [courseId]
    UpdateProgressBody:
      type: object
      properties:
        progress:
          type: integer
      required: [progress]
...