# lib/api-zod/src/generated/types/courseWithLessons.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from "./lesson";

export interface CourseWithLessons {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  createdBy: number;
  createdAt: string;
  updatedAt: string;
  lessonCount: number;
  enrollmentCount: number;
  lessons: Lesson[];
  isEnrolled: boolean;
  /** @nullable */
  progress?: number | null;
}