# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Enrollment {
  id: number;
  userId: number;
  courseId: number;
  progress: number;
  createdAt: string;
}