# lib/api-zod/src/generated/types/socialLinks.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialLinks {
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  tiktok?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  linkedin?: string | null;
}