# lib/api-zod/src/generated/types/updateCourseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCourseBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
}